test(corpus): keep cleanup behind platform facade
