+++ crates/unica-coder/tests/platform/format_8_3_27_xml_corpus.rs
@@ -1,5 +1,55 @@
 use std::path::Path;
 
+#[cfg(windows)]
+use std::io;
+#[cfg(windows)]
+use std::thread;
+#[cfg(windows)]
+use std::time::{Duration, Instant};
+
+#[cfg(windows)]
+fn is_windows_sharing_violation(error: &io::Error) -> bool {
+    const WINDOWS_SHARING_VIOLATION: i32 = 32;
+    error.raw_os_error() == Some(WINDOWS_SHARING_VIOLATION)
+}
+
+#[cfg(windows)]
+pub(super) fn remove_temp_tree(root: &Path) {
+    const RETRY_BUDGET: Duration = Duration::from_secs(15);
+    const RETRY_DELAY: Duration = Duration::from_millis(50);
+
+    let deadline = Instant::now() + RETRY_BUDGET;
+    loop {
+        match std::fs::remove_dir_all(root) {
+            Ok(()) => return,
+            Err(error) if error.kind() == io::ErrorKind::NotFound => return,
+            Err(error) if is_windows_sharing_violation(&error) && Instant::now() < deadline => {
+                thread::sleep(RETRY_DELAY);
+            }
+            Err(error) => panic!("cannot remove temporary tree {}: {error}", root.display()),
+        }
+    }
+}
+
+#[cfg(not(windows))]
+pub(super) fn remove_temp_tree(root: &Path) {
+    if let Err(error) = std::fs::remove_dir_all(root) {
+        assert_eq!(
+            error.kind(),
+            std::io::ErrorKind::NotFound,
+            "cannot remove temporary tree {}: {error}",
+            root.display()
+        );
+    }
+}
+
+#[cfg(windows)]
+#[test]
+fn windows_corpus_cleanup_recognizes_the_observed_sharing_violation() {
+    let error = io::Error::from_raw_os_error(32);
+    assert!(is_windows_sharing_violation(&error));
+}
+
 #[cfg(unix)]
 pub(super) fn require_single_link(path: &Path) -> Result<(), String> {
     use std::os::unix::fs::MetadataExt;
